requirements.txt
fastapi[all]==0.85.0
sqlalchemy==1.4.42
asyncpg==0.26.0
PyJWT==2.5.0
bcrypt==4.0.1
flake8==5.0.4
psycopg2-binary==2.9.4
pytest==7.2.0
Pillow==9.4.0
